Fix noise memory leak in examples/base64.c
Obviously we don't care if examples/base64.c leaks memory before
exiting, but things like that are noise that make it hard to tell if
there's a real memory-leak problem somewhere else; it also sets a bad
example for code that someone might write modeled on the example.
This patch ensures that the example frees its allocations as it
should.