avoid a segfault when h_act_flatten() is called on a null token
also includes commented-out tests for functionality that i discovered we never
provided, namely position and length information at the granularity of
HParsedTokens. at the moment, these fields always seem to be 0. we should
probably fix that or get rid of them. in the latter case it might make sense
to move this information to HParseResult, so the user can get access to it via
h_action() as needed.