"taint" memoized follow sets that came from intermediate results
tentative commit. this should solve issue 92, but isn't working, yet.

the idea is to taint any memoized follow set that hit a recursive loop,
directly or indirectly. the recursive calls (h_follow_rec) are allowed to use
these intermediates, but propagate the taint. the top-level routine (h_follow)
places the original taint on its result and may remove it at the end. it will
not return memoized results that are tainted (but compute them afresh).