add clang to tools if requested (or on Darwin)
clang appears to require -fPIC for shared objects on most platforms. the
'clang' tool adds the flag when appropriate. however, scons never selects the
clang tool automatically!

so we add the tool ourselves if CC=clang is set or generally on Darwin.

ideally, scons would know and handle all this, but at the time of writing
(scons v4.3.0), it does not.