dummy support for iterative parsing in packrat
A very first step. This implementation still expects all input in a single
chunk but allows the use of the iterative API. If the parser attempts to read
past the first chunk, the parse fails.

Contains some comments for next steps towards full support.

Adds tests for the single-chunk case.