remove an unreachable case
Replace it with an assert. This case could never occur because it tests
precisely the loop condition and there are no break statements in the loop.

This was the only use of the 'err' label, so that can go. The code under it
remains the fall-through case for 'err0', i.e. the actual error (parse failure)
case.