add (naive) multi-chunk operation for packrat
The naive implementation concatenates all input and blindly re-runs the full
parse on every chunk. Keeping state between chunks is for later.

Note: The iterative API expects us to always consume an entire input chunk
when we suspend, even if packrat later backtracks into it. We will produce
the correct parse result and accurately consume from a final chunk, but all
earlier chunks will be reported as fully consumed and as being part of the
HParseResult in terms of its bit_length field.